Correct region implementation and add tests
