+++ LoRaEngine/test/LoRaWanNetworkServer.Test/MessageProcessor_End2End_NoDep_Processing_Tests.cs
@@ -581,7 +581,7 @@ namespace LoRaWan.NetworkServer.Test
             Assert.NotNull(confirmedMessageResult.Txpk);
 
             // validates txpk according to eu region
-            Assert.Equal(RegionFactory.CreateEU868Region().GetDownstreamChannel(rxpk), confirmedMessageResult.Txpk.Freq);
+            Assert.Equal(RegionFactory.CreateEU868Region().GetDownstreamChannelFrequency(rxpk), confirmedMessageResult.Txpk.Freq);
             Assert.Equal("4/5", confirmedMessageResult.Txpk.Codr);
             Assert.False(confirmedMessageResult.Txpk.Imme);
             Assert.True(confirmedMessageResult.Txpk.Ipol);
@@ -745,7 +745,7 @@ namespace LoRaWan.NetworkServer.Test
             var confirmedMessageResult = await messageProcessor.ProcessMessageAsync(rxpk);
             Assert.NotNull(confirmedMessageResult);
             Assert.Equal(0U, confirmedMessageResult.Txpk.Rfch);
-            Assert.Equal(RegionFactory.CreateEU868Region().GetDownstreamChannel(rxpk), confirmedMessageResult.Txpk.Freq);
+            Assert.Equal(RegionFactory.CreateEU868Region().GetDownstreamChannelFrequency(rxpk), confirmedMessageResult.Txpk.Freq);
             Assert.Equal("4/5", confirmedMessageResult.Txpk.Codr);
             Assert.False(confirmedMessageResult.Txpk.Imme);
             Assert.True(confirmedMessageResult.Txpk.Ipol);
